requirements.txt
discord.py
asyncio
requests
humanfriendly
pymongo
python-dotenv
googletrans==4.0.0-rc1
nacl
